move models into persistence package
