+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/networking/SearchQuery.kt
@@ -2,7 +2,7 @@ package app.omnivore.omnivore.networking
 
 import app.omnivore.omnivore.graphql.generated.SearchQuery
 import app.omnivore.omnivore.graphql.generated.TypeaheadSearchQuery
-import app.omnivore.omnivore.models.LinkedItem
+import app.omnivore.omnivore.persistence.entities.LinkedItem
 import com.apollographql.apollo3.api.Optional
 
 data class SearchQueryResponse(
