+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/networking/ViewerQuery.kt
@@ -1,7 +1,7 @@
 package app.omnivore.omnivore.networking
 
 import app.omnivore.omnivore.graphql.generated.ViewerQuery
-import app.omnivore.omnivore.models.Viewer
+import app.omnivore.omnivore.persistence.entities.Viewer
 
 suspend fun Networker.viewer(): Viewer? {
   try {
